requirements.txt
psycopg2-binary==2.9.9
PyQt5==5.15.9